+++ packages/config/src/validate/lifecycle.js
@@ -1,0 +1,26 @@
+const LIFECYCLE = [
+  'init',
+  'preGetCache',
+  'getCache',
+  'postGetCache',
+  'preInstall',
+  'install',
+  'postInstall',
+  'preBuild',
+  'build',
+  'preFunctionsBuild',
+  'functionsBuild',
+  'postFunctionsBuild',
+  'postBuild',
+  'prePackage',
+  'package',
+  'postPackage',
+  'preDeploy',
+  'preSaveCache',
+  'saveCache',
+  'postSaveCache',
+  'finally',
+  'onError',
+]
+
+module.exports = { LIFECYCLE }
